Moved FigureExport class to figure_export.py out of Figure_to_Pdf.py
